Update README.md
This is a simple PHP + MySQL project I made for collecting student details during an industrial visit. It uses an HTML form that stores the data in a MySQL database using PHP. I used XAMPP to run it locally, and the SQL file is included so anyone can import and use it. This helped me learn how frontend and backend work together.